djangoProject/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg==3.1.18
psycopg2-binary==2.9.9